fix: show yes/no buttons on predict page mobile view
